refactor(router): split gestures into composable components
Pulls iOS edge swipe-back out of the `StackTransition` trait and
into a first-class `IosSwipeBack` component that users mount as a
child of `StackLayout`. The transition trait becomes pure visual
description (`animate` / `slot_style` / `foreground` only); gestures
and other layout-level interactions ride on the
`StackLayoutHandle` the layout now publishes to context.

The previous `StackTransition::setup(ctx: TransitionContext)` hook
was awkward in two ways:

1. The single trait method tried to cover all interactive behaviour
   (touch gestures, predictive back, hardware keys) under one hood.
   That left transition implementations responsible for input loops
   that have nothing to do with how the screen looks.
2. `TransitionContext` was a curated closure-bag — discoverable but
   inconsistent with the rest of the API. Users already know
   `router::<R>()` for stack ops; a separate `ctx.back()` next to it
   meant two ways to express the same thing.

The refactor:

- **`StackTransition` trait shrinks** to `animate` + `slot_style`
  (default) + `foreground` (default). Three methods, all pure
  visual concerns. Writing a new transition is now just "describe
  the look".
- **`StackLayoutHandle` lives in the layout** and goes into context
  via `provide_context`. It exposes the layout-internal primitives
  child components need (`container`, `current_wrapper`,
  `mount_preview`, `dispose_preview`, `commit_preview_and_back`).
- **`IosSwipeBack` is a regular `#[component]`** under
  `whisker_router::gestures`. It pulls the handle from context, uses
  `router::<R>()` directly for the back operation it needs, binds
  touch handlers in `on_mount`, and renders no DOM of its own.
- **`StackLayout` takes a `Children` slot** so users can compose
  gestures into it: `StackLayout { IosSwipeBack() }`.

Mixing and matching now composes naturally. `Fade` transition with
`IosSwipeBack` is a sentence the type system will let you write
(though the gesture's drag pose stays iOS-flavored — visual
consistency is on the author of the composition). Android system
back / predictive back can land as separate components in the same
`gestures` module without touching the trait.

`pose` returns to being a private helper of `IosSlide` (shared with
`IosSwipeBack` via a `pub(crate)` function), `TransitionContext` is
gone, and the demo grows one extra line:

```rust
StackLayout(render: render.clone()) {
    IosSwipeBack()
}
```
